Add option to use Pytorch
Model runner currently only supporst `keras`. We would like to use
either keras or pytorch here.

Current behavior was essentially checking the name of the framework
and importing a module with a correctly formatted name. This change
essentially does the same thing, adding a second conditional
statment in `import_pkg()`.